+++ _lp/core/lib/core.function.php
@@ -5,11 +5,11 @@ function lp_version()
 	return '3.1.0';
 }
 
-function transcribe($aList, $aIsTopLevel = true) 
+function transcribe($aList, $aIsTopLevel = true)
 {
    $gpcList = array();
    $isMagic = get_magic_quotes_gpc();
-  
+
    foreach ($aList as $key => $value) {
        if (is_array($value)) {
            $decodedKey = ($isMagic && !$aIsTopLevel)?stripslashes($key):$key;
@@ -23,14 +23,14 @@ function transcribe($aList, $aIsTopLevel = true)
    return $gpcList;
 }
 
-$_GET = transcribe( $_GET ); 
-$_POST = transcribe( $_POST ); 
+$_GET = transcribe( $_GET );
+$_POST = transcribe( $_POST );
 $_REQUEST = transcribe( $_REQUEST );
 
 
-function v( $str )
+function v( $str, $default = false )
 {
-	return isset( $_REQUEST[$str] ) ? $_REQUEST[$str] : false;
+	return isset( $_REQUEST[$str] ) ? $_REQUEST[$str] : $default;
 }
 
 function z( $str )
@@ -45,7 +45,7 @@ function c( $str )
 
 function g( $str )
 {
-	return isset( $GLOBALS[$str] ) ? $GLOBALS[$str] : false;	
+	return isset( $GLOBALS[$str] ) ? $GLOBALS[$str] : false;
 }
 
 function t( $str )
@@ -77,7 +77,7 @@ function wintval( $string )
 	{
 		if( is_numeric( $v ) ) $ret .= intval( $v );
 	}
-	
+
 	return $ret;
 }
 
@@ -130,10 +130,10 @@ function render( $data = NULL , $layout = NULL , $sharp = 'default' )
 			$layout = 'web';
 		}
 	}
-	
+
 	$GLOBALS['layout'] = $layout;
 	$GLOBALS['sharp'] = $sharp;
-	
+
 	$layout_file = AROOT . 'view/layout/' . $layout . '/' . $sharp . '.tpl.html';
 	if( file_exists( $layout_file ) )
 	{
@@ -147,7 +147,7 @@ function render( $data = NULL , $layout = NULL , $sharp = 'default' )
 		{
 			@extract( $data );
 			require( $layout_file );
-		}	
+		}
 	}
 }
 
@@ -159,7 +159,20 @@ function render_html( $data , $tpl )
 	$content = ob_get_contents();
 	ob_end_clean();
 	return $content;
-	// 
+	//
+}
+
+function render_ajax( $data )
+{
+   if( !headers_sent() )
+    {
+        header("Content-Type:text/json;charset=utf-8");
+        header("Expires: Thu, 01 Jan 1970 00:00:01 GMT");
+        header("Cache-Control: no-cache, must-revalidate");
+        header("Pragma: no-cache");
+    }
+
+    echo json_encode( $data );
 }
 
 function ajax_echo( $info )
@@ -171,7 +184,7 @@ function ajax_echo( $info )
 		header("Cache-Control: no-cache, must-revalidate");
 		header("Pragma: no-cache");
 	}
-	
+
 	echo $info;
 }
 
@@ -182,12 +195,12 @@ function info_page( $info , $title = '系统消息' )
 		$layout = 'ajax';
 	else
 		$layout = 'web';
-	
+
 	$data['top_title'] = $data['title'] = $title;
 	$data['info'] = $info;
-	
+
 	render( $data , $layout , 'info' );
-	
+
 }
 
 function smart_box( $info )
@@ -199,7 +212,7 @@ function smart_box( $info )
 		$array['error_message'] = $info;
 		return ajax_echo( json_encode( $array ));
 	}
-	elseif( is_ajax_request() ) 
+	elseif( is_ajax_request() )
 	{
 		return ajax_echo( $info );
 	}
@@ -209,25 +222,25 @@ function smart_box( $info )
 	}
 }
 
-if (!function_exists('apache_request_headers')) 
-{ 
+if (!function_exists('apache_request_headers'))
+{
 	function apache_request_headers()
-	{ 
+	{
 		foreach($_SERVER as $key=>$value)
-		{ 
+		{
 			if (substr($key,0,5)=="HTTP_")
-			{ 
-				$key=str_replace(" ","-",ucwords(strtolower(str_replace("_"," ",substr($key,5))))); 
-                    $out[$key]=$value; 
+			{
+				$key=str_replace(" ","-",ucwords(strtolower(str_replace("_"," ",substr($key,5)))));
+                    $out[$key]=$value;
 			}
 			else
-			{ 
-				$out[$key]=$value; 
+			{
+				$out[$key]=$value;
 			}
-       } 
-       
-	   return $out; 
-   } 
+       }
+
+	   return $out;
+   }
 }
 
 
@@ -255,21 +268,21 @@ function is_json_request()
 function is_mobile_request()
 {
     $_SERVER['ALL_HTTP'] = isset($_SERVER['ALL_HTTP']) ? $_SERVER['ALL_HTTP'] : '';
- 
+
     $mobile_browser = '0';
- 
+
     if(preg_match('/(up.browser|up.link|mmp|symbian|smartphone|midp|wap|phone|iphone|ipad|ipod|android|xoom)/i', strtolower($_SERVER['HTTP_USER_AGENT'])))
         $mobile_browser++;
- 
+
     if((isset($_SERVER['HTTP_ACCEPT'])) and (strpos(strtolower($_SERVER['HTTP_ACCEPT']),'application/vnd.wap.xhtml+xml') !== false))
         $mobile_browser++;
- 
+
     if(isset($_SERVER['HTTP_X_WAP_PROFILE']))
         $mobile_browser++;
- 
+
     if(isset($_SERVER['HTTP_PROFILE']))
         $mobile_browser++;
- 
+
     $mobile_ua = strtolower(substr($_SERVER['HTTP_USER_AGENT'],0,4));
     $mobile_agents = array(
                         'w3c ','acs-','alav','alca','amoi','audi','avan','benq','bird','blac',
@@ -282,21 +295,21 @@ function is_mobile_request()
                         'tosh','tsm-','upg1','upsi','vk-v','voda','wap-','wapa','wapi','wapp',
                         'wapr','webc','winw','winw','xda','xda-'
                         );
- 
+
     if(in_array($mobile_ua, $mobile_agents))
         $mobile_browser++;
- 
+
     if(strpos(strtolower($_SERVER['ALL_HTTP']), 'operamini') !== false)
         $mobile_browser++;
- 
+
     // Pre-final check to reset everything if the user is on Windows
     if(strpos(strtolower($_SERVER['HTTP_USER_AGENT']), 'windows') !== false)
         $mobile_browser=0;
- 
+
     // But WP7 is also Windows, with a slightly different characteristic
     if(strpos(strtolower($_SERVER['HTTP_USER_AGENT']), 'windows phone') !== false)
         $mobile_browser++;
- 
+
     if($mobile_browser>0)
         return true;
     else
@@ -308,21 +321,21 @@ function uses( $m )
 	load( 'lib/' . basename($m)  );
 }
 
-function load( $file_path ) 
+function load( $file_path )
 {
 	$file = AROOT . $file_path;
 	if( file_exists( $file ) )
 	{
 		//echo $file;
 		require( $file );
-	
+
 	}
 	else
 	{
 		//echo CROOT . $file_path;
 		require( CROOT . $file_path );
 	}
-	
+
 }
 
 // ===========================================
@@ -339,7 +352,7 @@ if( defined('SAE_APPNAME') )
 else
 	include_once( CROOT .  'lib/db' . $dbfile_postfix );
 
-// i18n 
+// i18n
 if (!function_exists('__'))
 {
 	function __( $string , $data = null )
@@ -347,7 +360,7 @@ if (!function_exists('__'))
 		if( !isset($GLOBALS['i18n']) )
 		{
 			$c = c('default_language');
-			if( strlen($c) < 1 ) $c = 'zh_cn';	
+			if( strlen($c) < 1 ) $c = 'zh_cn';
 		}
 		else
 			$c = z(t($GLOBALS['i18n']));
@@ -363,29 +376,29 @@ if (!function_exists('__'))
 			else
 			$GLOBALS['i18n'] = 'zh_cn';
 		}
-		
-		
-		
-		
+
+
+
+
 		if( isset( $GLOBALS['language'][$GLOBALS['i18n']][$string] ) )
 			$to = $GLOBALS['language'][$GLOBALS['i18n']][$string];
 		else
 			$to = $string;
-		
+
 		if( $data == null )
 			return $to;
 		else
 		{
 			if( !is_array( $data ) ) $data = array( $data );
 			return vsprintf( $to , $data );
-		}	
-			
+		}
+
 	}
-} 	
+}
 
 // **************************************************************
 // * Plugins & hooks
-// ************************************************************** 
+// **************************************************************
 function add_filter( $tag , $function_to_add , $priority = 10 , $accepted_args_num = 1 )
 {
     return add_hook( $tag , $function_to_add , $priority , $accepted_args_num );
@@ -454,16 +467,16 @@ function remove_hook( $tag , $priority = null )
     if( is_null($priority) ) unset( $GLOBALS['TTHOOK'][$tag] );
     else unset( $GLOBALS['TTHOOK'][$tag][$priority] );
 }
-// This function is based on wordpress  
+// This function is based on wordpress
 // from  https://raw.github.com/WordPress/WordPress/master/wp-includes/plugin.php
 // requere php5.2+
 
-function build_hook_id( $tag , $function ) 
+function build_hook_id( $tag , $function )
 {
     if ( is_string($function) )
         return $function;
 
-    if ( is_object($function) ) 
+    if ( is_object($function) )
     {
         // Closures are currently implemented as objects
         $function = array( $function, '' );
@@ -473,10 +486,10 @@ function build_hook_id( $tag , $function )
         $function = (array) $function;
     }
 
-    if (is_object($function[0]) ) 
+    if (is_object($function[0]) )
     {
         // Object Class Calling
-        if ( function_exists('spl_object_hash') ) 
+        if ( function_exists('spl_object_hash') )
         {
             return spl_object_hash($function[0]) . $function[1];
         }
@@ -518,11 +531,11 @@ function not_empty( $value )
 {
 	if( strlen( $value ) < 1 ) return false;
 	else return $value ;
-}	
+}
 
 function is_mail( $value )
 {
-	if( filter_var($value, FILTER_VALIDATE_EMAIL) === false ) 
+	if( filter_var($value, FILTER_VALIDATE_EMAIL) === false )
 		return false;
 	else
 		return $value;