+++ config/app.config.php
@@ -1,5 +1,5 @@
 <?php
-$GLOBALS['config']['site_name'] = 'LazyPHP3';
+$GLOBALS['config']['site_name'] = 'Mugen';
 $GLOBALS['config']['site_domain'] = 'lazyphp3.sinaapp.com';
 $GLOBALS['config']['site_icon'] = image('lazyphplogowithoutname.128.png');