+++ config/bootstrap.php
@@ -1,0 +1,3 @@
+<?php
+
+uses('user.class.php');