+++ index.php
@@ -4,7 +4,7 @@
 /****  load lp framework  ***/
 define( 'DS' , DIRECTORY_SEPARATOR );
 define( 'AROOT' , dirname( __FILE__ ) . DS  );
-
-//ini_set('include_path', dirname( __FILE__ ) . DS .'_lp' ); 
+session_start();
+//ini_set('include_path', dirname( __FILE__ ) . DS .'_lp' );
 include_once( '_lp'.DS .'lp.init.php' );
 /**** lp framework init finished ***/